LPS-190102 Allow users to sort in the fragment
